requirements.txt
# Основные зависимости для Telegram бота
python-telegram-bot==21.0.1
openai>=1.30.0
python-dotenv==1.0.1
aiosqlite==0.19.0
pillow==10.2.0

# Зависимости для СБП Payment Service
flask==3.0.0
requests==2.31.0
qrcode[pil]==7.4.2
aiohttp==3.9.1
